Adjust timing a debounce to prevent double digits on AC125A.
While testing with the AC125A, I found that sometimes, a single DTMF digit
would generate two relay operations.  This seemed to be related to how long the
Touch-Tone key was pressed.  I believe this is related to noise on the A and B
leads when the PA-PE relays switch.  To avoid this issue, clear any
pending DTMF digit that may have been queued by the ISR after actuating
the relays.